Got compound blocks to work.
You can look in helloWorldMain.mm for an example of how to use them. It basically comes down to: define a bunch of blocks and give them positions that will act as their offsets. Then make a compound block that takes in an array of these blocks.

Note that compound blocks cannot yet be used inside other compound blocks. I'll work on adding that in later.