Basic Wind Working.
I got a basic but realistic wind simulation working. It uses the neato box2d world raycast function. Right now, the wind just pops in automatically after the player runs out of objects. The test level right now somewhat tests out the wind. I'll be working next on getting the wind to appear for a period of time, pause before coming in, and act as part of the "disaster" class.