Added Level editor skeleton
Hey guys, I've started working on a level editor in java. I might be able to have it finished by the end of the weekend. I chose java primarily so that eveyone can use it. Right now, you should be able to just compile the project (I did it in Eclipse) to run it. I'll upload an executable binary once it's complete.

-Alec