Added the EarthQuake disaster and scrapped the java level editor.
The earthquake disaster is now in. I also gave up on making a level editor from scratch because it became apparent very quickly that doing so would take a non-trivial amount of time. For producing about five levels, it should be fairly simple to just type out the xml by hand. It's been working so far for my test levels.