Fixed a lot of memory management bugs with the level parser.
I went ahead and fixed the bugs Sarah was talking about with the memory management in the level parser. There were some in there that might've eventually bitten us in the ass in terms of memory leaks. I also simplified a few things with the parser to make it easier to use.